feat: update Button component and refactor Header layout
